feat(database): Add get_database_types function
This commit adds a new function called `get_database_types` to the `PostgresDatabaseType` struct. The function returns a vector of all the database types for the given data type category. It takes in references to a vector of data types for the given data type category and a reference to the data type category. The function has been tested using an example in the code. The test case verifies that the returned vector contains the expected number of database types and that the types are correct for the given category.