Refactor migration creation and execution code
In this commit, the code for creating and executing database migrations has been refactored. This includes changes in the `create_migration` function and the `run_migration` function.

In the `create_migration` function, the code for creating the `up.sql` and `down.sql` files has been reorganized. Additionally, unnecessary variables have been removed. The up.sql contents are now obtained from the `column_loop` function, and the down.sql contents are set to `DROP TABLE` statement for the table name.

In the `run_migration` function, the code for retrieving the database has been updated to use the `Database::get_database_from_rustyroad_toml` method. The `println!` statement for displaying the database type has been modified.

These changes improve the code structure and simplify the migration creation and execution process.