Fix typo in README
The commit message has been updated to correct a typo in the README file. The change is intended to clarify that the "main" branch should not be used for production purposes, instead of the incorrect mention of the "master" branch. The README file also includes information about the release candidate version and its fixes for the database connection strings.