feat(sql_writer): Add function to write SQL to file
This commit adds a new function `write_to_sql` to the SQL Writer module. The function takes in two arguments: the name of the file and the SQL code to be written. It checks if the SQL code is valid before creating a template and writing it to the file.

Example usage:
```
use crate::writers::general::sql_writer::write_to_sql;

write_to_sql("test.sql", "SELECT * FROM test;".to_string());
```
The result is the SQL code being written to the file.