feat: Add new controller creation function
This commit introduces a new function in the code that allows the creation of a new controller. The function checks if the current directory is a RustyRoad project and prompts the user to create a new directory for the controller if it doesn't exist. It also adds the controller to the corresponding file and creates a new file with the controller's name.