Update README with instructions to set `POSTGRES_LIB_PATH` environment variable
The existing README file has been updated to include instructions to set the `POSTGRES_LIB_PATH` environment variable on Windows. This environment variable should point to the directory containing `libpq.lib`, which is necessary for the project to find the libpq library. The instructions guide the user through the process of setting up the environment variable using the Windows user interface. Please note that the user should replace the example path with their actual PostgreSQL installation path.