feat: Add create controller functionality
This commit adds the ability to create new controllers for a rustyroad project. Users are prompted to enter the name of the controller and specify whether they want to add it to an existing folder or create a new one. If adding to an existing folder, the controller is added to the appropriate file. If creating a new folder, the necessary directories and files are created.