feat: Add create_update_model function
This commit adds the `create_update_model` function that is used to create an update model. The function checks if the current directory is a Rustyroad project by looking for the `rustyroad.toml` file. If the project does not have the file, it returns an error. It also looks for an existing model with the same name and creates a file if it doesn't exist. In addition, it writes the update model to the file.